+++ minecraft/src/net/minecraft/block/NetherPortalBlock.java
@@ -120,9 +120,9 @@ public class NetherPortalBlock extends Block {
 		}
 
 		for (int i = 0; i < 4; i++) {
-			double d = (double)pos.getX() + (double)random.nextFloat();
-			double e = (double)pos.getY() + (double)random.nextFloat();
-			double f = (double)pos.getZ() + (double)random.nextFloat();
+			double d = (double)pos.getX() + random.nextDouble();
+			double e = (double)pos.getY() + random.nextDouble();
+			double f = (double)pos.getZ() + random.nextDouble();
 			double g = ((double)random.nextFloat() - 0.5) * 0.5;
 			double h = ((double)random.nextFloat() - 0.5) * 0.5;
 			double j = ((double)random.nextFloat() - 0.5) * 0.5;
