+++ minecraft/src/net/minecraft/client/sound/ElytraSoundInstance.java
@@ -24,9 +24,9 @@ public class ElytraSoundInstance extends MovingSoundInstance {
 	public void tick() {
 		this.tickCount++;
 		if (!this.player.removed && (this.tickCount <= 20 || this.player.isFallFlying())) {
-			this.x = (float)this.player.getX();
-			this.y = (float)this.player.getY();
-			this.z = (float)this.player.getZ();
+			this.x = (double)((float)this.player.getX());
+			this.y = (double)((float)this.player.getY());
+			this.z = (double)((float)this.player.getZ());
 			float f = (float)this.player.getVelocity().lengthSquared();
 			if ((double)f >= 1.0E-7) {
 				this.volume = MathHelper.clamp(f / 4.0F, 0.0F, 1.0F);
