+++ minecraft/src/net/minecraft/entity/ai/goal/FollowOwnerGoal.java
@@ -114,7 +114,7 @@ public class FollowOwnerGoal extends Goal {
 		} else if (!this.canTeleportTo(new BlockPos(x, y, z))) {
 			return false;
 		} else {
-			this.tameable.refreshPositionAndAngles((double)((float)x + 0.5F), (double)y, (double)((float)z + 0.5F), this.tameable.yaw, this.tameable.pitch);
+			this.tameable.refreshPositionAndAngles((double)x + 0.5, (double)y, (double)z + 0.5, this.tameable.yaw, this.tameable.pitch);
 			this.navigation.stop();
 			return true;
 		}
